fix(eslint): strongly type recommended rules to match rules
